backend/requirements.txt
asgiref==3.9.1
Django==5.2.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
gunicorn==23.0.0
packaging==25.0
psycopg2==2.9.10
sqlparse==0.5.3
typing_extensions==4.15.0
